+++ src/interfaces/ApiPluginInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace PwJsonApi\Plugins;
+
+use PwJsonApi\{Api, Service, Endpoint};
+
+interface ApiPluginInterface
+{
+  public function init(Api|Service|Endpoint $context): static;
+}